Revert "refactor(app): rename error recovery "hooks" dir to "utils""
This reverts commit 46cb0e44dd64ed49803b4f482dbeae03b6114d30.
