+++ app/src/organisms/ErrorRecoveryFlows/hooks/__tests__/getFailedCommandPipetteInfo.test.ts
